+++ CHANGELOG.md
@@ -25,7 +25,10 @@ Format follows [Keep a Changelog](https://keepachangelog.com/).
 
 
 
+
 ## [Unreleased]
+
+## [0.3.173] - 2026-08-29
 
 ## [0.3.172] - 2026-08-29